// src/interfaces/common/product-damage.interface.ts
export interface ProductDamage {
  _id: string;
  product?: any;
  quantity: number;
  dateString?: string;
  shop?: any;
  month?: number;
  year?: number;
  updateTime?: any;
  category?: any;
  vendor?: any;
  previousQuantity?: number;
  updatedQuantity?: number;
  imei?: any;
  note?: string;
  salesman?: any;
  createTime?: any;
  createdAt?: Date;
  updatedAt?: Date;
  select?: boolean;
}